vboy.c: Fixed crash regression (nw) The system hangs when I tried a game though
git-svn-id: svn://messdev.no-ip.org/mess@14923 9849079f-6229-0410-a07f-84406ce305ba

+++ src/mess/drivers/vboy.c
@@ -76,10 +76,10 @@ public:
 	DECLARE_READ16_MEMBER(vboy_bgmap_r);
 	UINT16 *m_font;
 	UINT16 *m_bgmap;
-	required_shared_ptr<UINT16> m_l_frame_0;
-	required_shared_ptr<UINT16> m_l_frame_1;
-	required_shared_ptr<UINT16> m_r_frame_0;
-	required_shared_ptr<UINT16> m_r_frame_1;
+	required_shared_ptr<UINT32> m_l_frame_0;
+	required_shared_ptr<UINT32> m_l_frame_1;
+	required_shared_ptr<UINT32> m_r_frame_0;
+	required_shared_ptr<UINT32> m_r_frame_1;
 	UINT16 *m_world;
 	UINT16 *m_columntab1;
 	UINT16 *m_columntab2;
